requirements.txt
asgiref
certifi
crispy-bootstrap5
distlib
dj-database-url
Django
django-crispy-forms
django-email
django-environ
django-filter
djangorestframework
EasyProcess
entrypoint2
filelock
Markdown
MouseInfo
mss
numpy
opencv-contrib-python
Pillow
pipenv
platformdirs
psycopg2
psycopg2-binary
PyAutoGUI
PyGetWindow
PyMsgBox
pyperclip
PyRect
pyscreenshot
PyScreeze
pytweening
pytz
sqlparse
typing_extensions
tzdata
virtualenv  